feat(header,switch): add 'switch' component and toggle theme
